requirements.txt
webdriver_manager==3.4.2
google_auth_oauthlib==0.4.5
requests==2.26.0
selenium==3.141.0
beautifulsoup4==4.9.3
google_api_python_client==2.15.0
Pillow==8.3.1
protobuf==3.17.3
